+++ modules/dsp/chowdsp_filters/chowdsp_filters.h
@@ -49,3 +49,4 @@ BEGIN_JUCE_MODULE_DECLARATION
 #include "Other/chowdsp_WernerFilter.h"
 #include "Other/chowdsp_ARPFilter.h"
 #include "Other/chowdsp_LinkwitzRileyFilter.h"
+#include "Other/chowdsp_ThreeWayCrossoverFilter.h"